Added support for subvolume snap shotting
In the current master branch btrfs-snap will not create a backup of a nested subvolume. The proposed changes add support for subvolumes by querying for subvolumes located in the specified mountpoint. If `btrfs subvolume list ${mp$}` returns an entry which matches the end of `${mp}` it is a valid subvolume.